ui: Show software selection menu before doing system audit, for casual browsing.
