+++ src/frontend/mame/ui/selsoft.cpp
@@ -468,8 +468,17 @@ void menu_select_software::build_software_list()
 void menu_select_software::inkey_select(const event *menu_event)
 {
 	ui_software_info *ui_swinfo = (ui_software_info *)menu_event->itemref;
+	driver_enumerator drivlist(machine().options(), *ui_swinfo->driver);
+	media_auditor auditor(drivlist);
+	drivlist.next();
 
-	if (ui_swinfo->startempty == 1)
+	// audit the system ROMs first to see if we're going to work
+	media_auditor::summary const sysaudit = auditor.audit_media(AUDIT_VALIDATE_FAST);
+	if (sysaudit != media_auditor::CORRECT && sysaudit != media_auditor::BEST_AVAILABLE && sysaudit != media_auditor::NONE_NEEDED)
+	{
+		set_error(reset_options::REMEMBER_REF, make_audit_fail_text(media_auditor::NOTFOUND != sysaudit, auditor));
+	}
+	else if (ui_swinfo->startempty == 1)
 	{
 		if (!select_bios(*ui_swinfo->driver, true))
 		{
@@ -479,16 +488,13 @@ void menu_select_software::inkey_select(const event *menu_event)
 	}
 	else
 	{
-		// first validate
-		driver_enumerator drivlist(machine().options(), *ui_swinfo->driver);
-		media_auditor auditor(drivlist);
-		drivlist.next();
+		// first audit the software
 		software_list_device *swlist = software_list_device::find_by_name(*drivlist.config(), ui_swinfo->listname);
 		const software_info *swinfo = swlist->find(ui_swinfo->shortname);
 
-		media_auditor::summary const summary = auditor.audit_software(*swlist, *swinfo, AUDIT_VALIDATE_FAST);
+		media_auditor::summary const swaudit = auditor.audit_software(*swlist, *swinfo, AUDIT_VALIDATE_FAST);
 
-		if (summary == media_auditor::CORRECT || summary == media_auditor::BEST_AVAILABLE || summary == media_auditor::NONE_NEEDED)
+		if (swaudit == media_auditor::CORRECT || swaudit == media_auditor::BEST_AVAILABLE || swaudit == media_auditor::NONE_NEEDED)
 		{
 			if (!select_bios(*ui_swinfo, false) && !select_part(*swinfo, *ui_swinfo))
 			{
@@ -500,11 +506,11 @@ void menu_select_software::inkey_select(const event *menu_event)
 		{
 			// otherwise, display an error
 			std::ostringstream str;
-			str << _("The selected software is missing one or more required files. Please select a different software.\n\n");
-			if (media_auditor::NOTFOUND != summary)
+			str << _("The selected software is missing one or more required files. Please select a different software item.\n\n");
+			if (media_auditor::NOTFOUND != swaudit)
 			{
 				auditor.summarize(nullptr, &str);
-				str << "\n";
+				str << '\n';
 			}
 			str << _("Press any key to continue."),
 			set_error(reset_options::REMEMBER_POSITION, str.str());
